--- src/hooks/UseScrollToSection.ts
@@ -1,10 +1,0 @@
-export const UseScrollToSection = () => {
-  const handleScrollToSection = (targetId: string) => {
-    const targetElement = document.getElementById(targetId);
-    if (targetElement) {
-      targetElement.scrollIntoView();
-    }
-  };
-
-  return { handleScrollToSection };
-};
