+++ src/hooks/useScrollToHash.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 
-export const UseScrollToHash = () => {
+export const useScrollToHash = () => {
   const location = useLocation();
 
   useEffect(() => {
