+++ src/hooks/useScrollToSection.ts
@@ -1,0 +1,11 @@
+export const useScrollToSection = () => {
+    const handleScrollToSection = (targetId: string) => {
+      const targetElement = document.getElementById(targetId);
+      if (targetElement) {
+        targetElement.scrollIntoView();
+      }
+    };
+  
+    return { handleScrollToSection };
+  };
+
